Add more chars in report message column
Extend the amount of characters allowed by default for the description column of the report to 500.

After using the functionality in production, it was clear that 100 was too short for most of the cases.

It is still limited, now to 500, for readability purposes, since the report is meant for business users to see.